# Inline ordered selections