--- a/README.md
+++ b/README.md
@@ -2,17 +2,17 @@
 
 **An (experimental) breadth-first GraphQL executor for Ruby**
 
-Depth-first execution resolves every object field descending down a response tree, while breadth-first visits every _selection position_ once with an aggregated set of objects. A breadth-first approach makes resolver overhead dramatically cheaper when resolvers only scale by the size of the request document rather than the size of the response.
+Depth-first execution resolves every object field descending down a response tree, while breadth-first visits every _selection position_ once with an aggregated set of objects. The breadth-first approach tends to be much faster due to fewer resolver calls and intermediary promises.
 
 ```shell
 graphql-ruby: 140002 resolvers
-    1.159 (± 0.0%) i/s  (862.55 ms/i) - 6.000 in 5.182856s
+   1.087 (± 0.0%) i/s  (919.76 ms/i) -  6.000 in  5.526807s
 graphql-cardinal 140002 resolvers
-    19.251 (±10.4%) i/s   (51.95 ms/i) - 95.000 in 5.007853s
+   21.314 (± 9.4%) i/s   (46.92 ms/i) -  108.000 in  5.095015s
 
 Comparison:
-graphql-cardinal 140002 resolvers:     19.3 i/s
-graphql-ruby: 140002 resolvers:     1.2 i/s - 16.60x  slower
+graphql-cardinal 140002 resolvers:   21.3 i/s
+graphql-ruby: 140002 resolvers:   1.1 i/s - 19.60x  slower
 ```
 
 ### Depth vs. Breadth
@@ -23,7 +23,7 @@ GraphQL requests have two dimensions: _depth_ and _breadth_. The depth dimension
 
 ### Depth-first execution
 
-Depth-first execution (the conventional GraphQL execution strategy) resolves every field in the response by descending down the selection tree of every object. This overhead scales linearly as the response size grows, and balloons quickly with added field tracing and instrumentation.
+Depth-first execution (the conventional GraphQL execution strategy) resolves every field in the response by descending down the selection tree of every object. This overhead scales as the response size grows, and balloons quickly with added field tracing and instrumentation.
 
 ![Depth](./images/depth-first.png)
 
@@ -41,7 +41,7 @@ Breadth-first then runs a single resolver per document selection, and coalesces 
 
 ![Breadth](./images/breadth-first.png)
 
-While bigger responses will always take longer to process, the workload is almost entirely your own business logic rather than GraphQL execution overhead. Other advantages:
+While bigger responses will always take longer to process, the workload is your own business logic with very little GraphQL execution overhead. Other advantages:
 
 * Eliminates the need for DataLoader promises, because resolvers are inherently batched.
 * Executes via flat queuing without deep recursion and huge call stacks.