--- a/lib/graphql/cardinal/executor/response_shape.rb
+++ b/lib/graphql/cardinal/executor/response_shape.rb
@@ -26,9 +26,7 @@ module GraphQL::Cardinal
             begin
               node_type = @query.get_field(parent_type, node.name).type
               named_type = node_type.unwrap
-
-              # delete and re-add to order result keys...
-              raw_value = raw_object.delete(field_name)
+              raw_value = raw_object[field_name]
 
               raw_object[field_name] = if raw_value.is_a?(ExecutionError)
                 # capture errors encountered in the response with proper path