Inlines selection order concerns with virtually no overhead, which speeds up the lazy execution benchmark by eliminating the associated shaping pass.

```
Before:
graphql-cardinal: 140002 lazy resolvers
    1.729 (± 0.0%) i/s  (578.29 ms/i) -      9.000 in   5.205185s

After:
graphql-cardinal: 140002 lazy resolvers
    2.025 (± 0.0%) i/s  (493.78 ms/i) -     11.000 in   5.433050s
```

This solution won't hold up to async field executions, but we can cross that bridge when we get there. For now, we have fast ordered keys.